Refactor: type import 위치 변경
