+++ app/member/commits/[sha]/memoir/components/CommitMemoir.tsx
@@ -8,9 +8,9 @@ import CreateMemoirLayout from "@/app/member/components/CreateMemoir/CreateMemoi
 import EditorForm from "@/app/member/components/CreateMemoir/EditorForm";
 import useExtractFilenames from "@/hooks/useExtractFileNames";
 
+import { useSummaryStore } from "@/store/AiSummaryStore";
 import { useRepoStore } from "@/store/repoStore";
 import { useParams } from "next/navigation";
-import { useSummaryStore } from "@/store/AiSummaryStore";
 
 import { CommitType } from "@/types/github/CommitType";
 import { useSession } from "next-auth/react";
